refactor: split gateway into a module and extract shared helpers
Move the gateway monolith into src/gateway/ (mod, worker, tests), extract
doctor into its own file, and share test scaffolding through test_support.
No behavior change; full test suite passes.
