+++ src/telegram.rs
@@ -151,7 +151,6 @@ impl Telegram {
                 .is_some_and(|id| self.allow_chat_ids.contains(&id))
     }
 
-    #[allow(dead_code)]
     pub fn allows_target(&self, chat_id: i64) -> bool {
         self.allow_user_ids.contains(&chat_id) || self.allow_chat_ids.contains(&chat_id)
     }
